requirements.txt
# Web & API
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.2
openai==1.58.1
python-dotenv==1.0.1

# RAG & AI
langchain==0.3.7
langchain-openai==0.3.2
langchain-community==0.3.7
faiss-cpu==1.8.0
numpy==1.24.4

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.2
pdf2image==1.17.0
PyMuPDF==1.24.1
python-multipart==0.0.9